+++ src/transform.jl
@@ -1,15 +1,3 @@
-"""[intended for internal use only]
-Convert all selected data columns to floats
-"""
-function _data_to_float!(e::Experiment)
-    # Make sure all values are numbers
-    @assert all( [x <: Number for x in eltype.(eachcol(getdata(e)))] )
-    # Convert each column to floats
-    for colname in names(getdata(e))
-        e.data[!,colname] = float.(e.data[:,colname])
-    end
-end
-
 """Approximate normal distribution"""
 logtransform(x) = log.(x .+ 1 .- minimum(x))
 
@@ -82,6 +70,7 @@ decorrelate(data::AbstractMatrix; ordercol = nothing, threshold = 0.8) =
     a given order 'ordercol' (defaults to left to right).
 """
 function decorrelate!(e::Experiment; ordercol = nothing, threshold = 0.8)
+    _assert_clean_data(e)
     e.selected_features = e.selected_features[decorrelate(getdata(e), 
                                                         ordercol=ordercol,
                                                         threshold=threshold)]
@@ -92,6 +81,7 @@ end
     with largest median absolute deviation.
 """
 function decorrelate_by_mad!(e::Experiment; threshold = 0.8)
+    _assert_clean_data(e)
     # Order features from biggest mad to smallest mad
     # When features have mad(reference) = 1, it means that we rank features 
     # by how more variable they are overall compared to the reference